BP+OPENEND+ENDING_SOUND+TIME+DELIMITER
-방정리 나래이션
-방정리 제한시간
-오프닝, 엔딩 나레이션
-엔딩 제자리 delimiter

+++ Assets/Scripts/Tutorial/AutoVoiceRecording_Ending.cs
@@ -70,19 +70,36 @@ public class AutoVoiceRecording_Ending : MonoBehaviour
         recordingNew.SetData(data, 0);
         recording = recordingNew;
         audioSource.clip = recording;
+        if(UserData.DataManager.GetInstance().FilePath_Folder!= null)
+        {
+            SavWav.Save(UserData.DataManager.GetInstance().FilePath_Folder + FileName + fWAV, recording);
+            //저장된 wav파일 mp3로 변환
+            WaveToMP3(UserData.DataManager.GetInstance().FilePath_Folder + FileName + fWAV,
+           UserData.DataManager.GetInstance().FilePath_Folder + FileName + fMP3);
 
-        //wav파일로 저장
-        //    SavWav.Save(UserData.DataManager.GetInstance().FilePath_Folder + FileName + fWAV, recording);
-        // 왕왕 중요하다 여기 수정해뒀으니까 꼭 다시 읽어봐라
-        SavWav.Save(@"C:\Desktop\" + "sound" + fWAV, recording);
-        //저장된 wav파일 mp3로 변환
-        WaveToMP3(@"C:\Desktop\" + "sound" + fWAV,
-       @"C:\Desktop\" + "sound" + fMP3);
+            //파일 저장 완료까지 대기
+            yield return new WaitUntil(() => File.Exists(UserData.DataManager.GetInstance().FilePath_Folder + FileName + fMP3));
 
-        //파일 저장 완료까지 대기
-        yield return new WaitUntil(() => File.Exists(@"C:\Desktop\" + "sound" + fMP3));  
+            File.Delete(UserData.DataManager.GetInstance().FilePath_Folder + FileName + fWAV);
 
-        File.Delete(@"C:\Desktop\" + "sound" + fWAV);
+        }
+        if(UserData.DataManager.GetInstance().FilePath_Folder == null)
+        {
+            //wav파일로 저장
+            //    SavWav.Save(UserData.DataManager.GetInstance().FilePath_Folder + FileName + fWAV, recording);
+            // 왕왕 중요하다 여기 수정해뒀으니까 꼭 다시 읽어봐라
+            SavWav.Save(@"C:\Desktop\" + "sound" + fWAV, recording);
+            //저장된 wav파일 mp3로 변환
+            WaveToMP3(@"C:\Desktop\" + "sound" + fWAV,
+           @"C:\Desktop\" + "sound" + fMP3);
+
+            //파일 저장 완료까지 대기
+            yield return new WaitUntil(() => File.Exists(@"C:\Desktop\" + "sound" + fMP3));
+
+            File.Delete(@"C:\Desktop\" + "sound" + fWAV);
+
+        }
+       
     }
